requirements.txt
django==4.2.4
django-rest-framework==3.14.0
djoser==2.2.0
ipython==8.14.0